backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
base58==2.1.1
bitcoinlib==0.7.1
cachetools==4.2.4
certifi==2025.1.31
charset-normalizer==3.4.1
coincurve==21.0.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cytoolz==1.0.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ecdsa==0.19.1
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
mysqlclient==2.2.7
NextVibeTestnet==0.1
numpy==2.2.4
packaging==24.2
parsimonious==0.10.0
pillow==11.1.0
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==8.0
qrcode-artistic==3.0.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
rfc3986==1.5.0
segno==1.6.1
six==1.17.0
sniffio==1.3.1
solana==0.27.0
solders==0.9.3
SQLAlchemy==2.0.39
sqlparse==0.5.3
toolz==1.0.0
tronpy==0.5.0
types-cachetools==4.2.10
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==10.4
